add private field to book do some code refactor

+++ backend/src/controllers/localBooks.controller.ts
@@ -1,9 +1,10 @@
 import { NextFunction, Response, Request } from 'express';
-import { getAllBooks, getBook, updateBook, deleteBook, createBook, getPaginatedBooks } from '../services/db/books.services';
+import { getAllBooks, getBook, updateBook, deleteBook, createBook, getPaginatedBooks, getBookFields } from '../services/db/books.services';
 import logger from '../helpers/logger';
-import { UnProcessableEntityError } from '../helpers/error';
+import { DontHaveAccessError, UnProcessableEntityError } from '../helpers/error';
 import { CreateBookRequest } from '../types/requests/books.request.type';
 import { getJsonBook } from '../utils/booktojson';
+import { IRequestWithUser } from '../types/requests/global.request.type';
 
 export const createBookAction = async (
   req: CreateBookRequest,
@@ -16,14 +17,14 @@ export const createBookAction = async (
 
   const user_id = req.user?.id;
   // const cover = `${req.file?.destination}${req.file?.filename}`; // обложка cover: req.file ? cover : null, 
-  const payload = {...req?.body, user_id, link: `${req.file?.destination}${req.file?.filename}`};
+  const payload = {...req?.body, private: req?.body.privat === 'true' ? true : false, user_id, link: `${req.file?.destination}${req.file?.filename}`};
 
   logger.info(`Create Book Action: { payload: ${JSON.stringify(payload)} }`);
 
   try {
     const {genres, tags, ...data} = payload;
 
-    const book = await createBook(data, genres.split(';'), tags.split(';'));
+    const book = await createBook(data, genres?.split(';'), tags?.split(';'));
     
     res.status(200).json(book);
   } catch (error) {
@@ -70,16 +71,17 @@ export const getPaginatedBooksAction = async (
 };
 
 export const getBookAction = async (
-  req: Request,
+  req: IRequestWithUser,
   res: Response,
   next: NextFunction
 ) => {
   const id = req.params.id;
+  const user_id = req.user?.id;
 
-  logger.info(`Get Book Action: { id: ${id} }`);
+  logger.info(`Get Book Action: { id: ${id}, user_id: ${user_id} }`);
 
   try {
-    const book = await getBook({id});
+    const book = await getBook({id}, user_id);
     const text = await getJsonBook(book.link);
 
     // const translatedChapter = await getTranslatedChapter(text.content[4])
@@ -92,21 +94,28 @@ export const getBookAction = async (
 };
 
 export const updateBookAction = async (
-  req: Request,
+  req: IRequestWithUser,
   res: Response,
   next: NextFunction
 ) => {
   const id = req.params.id;
   const payload = req.body;
+  const user_id = req.user?.id;
 
   logger.info(`Update Book Action: { id: ${id}, payload: ${JSON.stringify(payload)} }`);
 
   try {
     const {genres, tags, ...data} = payload;
 
-    const book = await updateBook({ id }, data, genres.split(';'), tags.split(';'));
-    
-    res.status(200).json(book);
+    const prevBook = await getBookFields({ id }, ['user_id']);
+
+    if (prevBook.user_id === user_id) {
+      const book = await updateBook({ id }, data, genres.split(';'), tags.split(';'));
+      
+      res.status(200).json(book);
+    } else {
+      throw new DontHaveAccessError();
+    }
   } catch (error) {
     logger.error('Update Book Action - Cannot update book', error);
     next(error);
@@ -114,17 +123,23 @@ export const updateBookAction = async (
 };
 
 export const deleteBookAction = async (
-  req: Request,
+  req: IRequestWithUser,
   res: Response,
   next: NextFunction
 ) => {
   const id = req.params.id;
+  const user_id = req.user?.id;
 
   logger.info(`Delete Book Action: { id: ${id} }`);
 
   try {    
-    // добавить проверку на пользователя
-    await deleteBook(id);
+    const book = await getBookFields({ id }, ['user_id']);
+
+    if (book.user_id === user_id) {
+      await deleteBook(id);
+    } else {
+      throw new DontHaveAccessError();
+    }
     
     res.status(200).json(id);
   } catch (error) {
